# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/outlets/{outletId}/timings" : {
      "put" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "updateOutletTimings",
        "parameters" : [ {
          "name" : "outletId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OutletTimingsUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outletId}/status" : {
      "put" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "updateOutletStatus",
        "parameters" : [ {
          "name" : "outletId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OutletStatusUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outlets/{outletId}/settings" : {
      "put" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "updateOutletSettings",
        "parameters" : [ {
          "name" : "outletId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OutletSettingsUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brands/{brandId}/outlets" : {
      "get" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "getOutletsByBrand",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListOutletDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "onboardOutlet",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OutletOnboardRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseOutletDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/restaurants/{id}" : {
      "get" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "getRestaurant",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNearbyRestaurantDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/restaurants/nearby" : {
      "get" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "getNearbyRestaurants",
        "parameters" : [ {
          "name" : "lat",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "lng",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "radius",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double",
            "default" : 5.0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListNearbyRestaurantDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/brands/{brandId}/outlets" : {
      "get" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "getBrandOutlets",
        "parameters" : [ {
          "name" : "brandId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "lat",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "lng",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number",
            "format" : "double"
          }
        }, {
          "name" : "radius",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number",
            "format" : "double",
            "default" : 5.0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListNearbyRestaurantDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/restaurants/all-with-location" : {
      "get" : {
        "tags" : [ "restaurant-outlet-controller" ],
        "operationId" : "getAllOutletsWithLocation",
        "parameters" : [ {
          "name" : "cityId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "pattern" : "^[A-Z][A-Z0-9_-]{0,63}$",
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseDtoNearbyRestaurantDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "OutletTimingsUpdateRequest" : {
        "required" : [ "timings" ],
        "type" : "object",
        "properties" : {
          "timings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimingRequest"
            }
          }
        }
      },
      "TimingRequest" : {
        "required" : [ "closingTime", "openingTime" ],
        "type" : "object",
        "properties" : {
          "openingTime" : {
            "type" : "string",
            "example" : "10:00:00"
          },
          "closingTime" : {
            "type" : "string",
            "example" : "22:00:00"
          }
        }
      },
      "ApiResponseVoid" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "OutletStatusUpdateRequest" : {
        "required" : [ "isActive" ],
        "type" : "object",
        "properties" : {
          "isActive" : {
            "type" : "boolean"
          }
        }
      },
      "OutletSettingsUpdateRequest" : {
        "required" : [ "defaultPrepTimeSeconds" ],
        "type" : "object",
        "properties" : {
          "defaultPrepTimeSeconds" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "OutletOnboardRequest" : {
        "required" : [ "fssaiLicenseNumber", "lat", "lng", "name", "timeZone", "timings" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "fssaiLicenseNumber" : {
            "type" : "string"
          },
          "lat" : {
            "type" : "number",
            "format" : "double"
          },
          "lng" : {
            "type" : "number",
            "format" : "double"
          },
          "cityId" : {
            "maxLength" : 64,
            "minLength" : 0,
            "pattern" : "^[A-Z][A-Z0-9_-]{0,63}$",
            "type" : "string"
          },
          "timings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimingRequest"
            }
          },
          "timeZone" : {
            "type" : "string"
          },
          "bannerUrl" : {
            "type" : "string"
          },
          "cuisine" : {
            "type" : "string"
          },
          "rating" : {
            "type" : "number",
            "format" : "double"
          },
          "reviewsCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deliveryTime" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deliveryFee" : {
            "type" : "number"
          },
          "tags" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseOutletDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/OutletDto"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "OutletDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "brandId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "fssaiLicenseNumber" : {
            "type" : "string"
          },
          "lat" : {
            "type" : "number",
            "format" : "double"
          },
          "lng" : {
            "type" : "number",
            "format" : "double"
          },
          "cityId" : {
            "type" : "string"
          },
          "bannerUrl" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "defaultPrepTimeSeconds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "cuisine" : {
            "type" : "string"
          },
          "rating" : {
            "type" : "number",
            "format" : "double"
          },
          "reviewsCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deliveryTime" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deliveryFee" : {
            "type" : "number"
          },
          "tags" : {
            "type" : "string"
          },
          "timings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OutletTimingDto"
            }
          },
          "timeZone" : {
            "type" : "string"
          }
        }
      },
      "OutletTimingDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "dayOfWeek" : {
            "type" : "integer",
            "format" : "int32"
          },
          "openingTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "closingTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          }
        }
      },
...
      "ApiResponseNearbyRestaurantDTO" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/NearbyRestaurantDTO"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponseListNearbyRestaurantDTO" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NearbyRestaurantDTO"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponseListOutletDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OutletDto"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponsePageResponseDtoNearbyRestaurantDTO" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseDtoNearbyRestaurantDTO"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}