# src/main/java/com/fooddelivery/restaurant/controller/RestaurantOutletController.java
package com.fooddelivery.restaurant.controller;

import com.fooddelivery.common.dto.ApiResponse;
import com.fooddelivery.restaurant.dto.*;
import com.fooddelivery.restaurant.entity.Brand;
import com.fooddelivery.restaurant.entity.Outlet;
import com.fooddelivery.restaurant.service.RestaurantOnboardingService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.security.access.prepost.PreAuthorize;
import java.util.UUID;
import java.util.List;
import java.util.Map;
import java.util.HashMap;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import org.springframework.validation.annotation.Validated;

import lombok.extern.slf4j.Slf4j;

@Slf4j
@RestController
@Validated
public class RestaurantOutletController {
    private final RestaurantOnboardingService onboardingService;

    public RestaurantOutletController(RestaurantOnboardingService onboardingService) {
        this.onboardingService = onboardingService;
    }

    private OutletTimingDto toOutletTimingDto(com.fooddelivery.restaurant.entity.OutletTiming timing) {
        return OutletTimingDto.builder().id(timing.getId()).openingTime(timing.getOpeningTime()).closingTime(timing.getClosingTime()).build();
    }

    private OutletDto toOutletDto(Outlet outlet) {
        List<OutletTimingDto> timingsDto = new java.util.ArrayList<>();
        if (outlet.getTimings() != null) {
            timingsDto = outlet.getTimings().stream().map(this::toOutletTimingDto).collect(java.util.stream.Collectors.toList());
        }
        return OutletDto.builder()
                .id(outlet.getId())
                .brandId(outlet.getBrandId())
                .name(outlet.getName())
                .fssaiLicenseNumber(outlet.getFssaiLicenseNumber())
                .lat(outlet.getLocation() != null ? outlet.getLocation().getY() : null)
                .lng(outlet.getLocation() != null ? outlet.getLocation().getX() : null)
                .cityId(outlet.getCityId())
                .bannerUrl(outlet.getBannerUrl())
                .isActive(outlet.getIsActive())
                .defaultPrepTimeSeconds(outlet.getDefaultPrepTimeSeconds())
                .cuisine(outlet.getCuisine())
                .rating(outlet.getRating())
                .reviewsCount(outlet.getReviewsCount())
                .deliveryTime(outlet.getDeliveryTime())
                .deliveryFee(outlet.getDeliveryFee())
                .tags(outlet.getTags())
                .timings(timingsDto)
                .timeZone(outlet.getTimeZone() != null ? outlet.getTimeZone().getId() : null)
                .build();
    }

    @GetMapping("/api/v1/outlets")
    @PreAuthorize("hasRole('RESTAURANT')")
    public ResponseEntity<ApiResponse<List<OutletDto>>> getOutlets(java.security.Principal principal) {
        List<OutletDto> outlets = onboardingService.getOutletsForUser(UUID.fromString(principal.getName()), com.fooddelivery.common.enums.OrganisationPermission.ORG_VIEW).stream().map(this::toOutletDto).collect(java.util.stream.Collectors.toList());
        return ResponseEntity.ok(ApiResponse.success(outlets, "Outlets retrieved successfully"));
    }

    // Phase 2: Outlet Onboarding
    @PostMapping("/api/v1/brands/{brandId}/outlets")
    @PreAuthorize("hasRole('RESTAURANT') and @restaurantAccess.onBrand(#brandId, authentication, T(com.fooddelivery.common.enums.OrganisationPermission).OUTLET_MANAGE)")
    public ResponseEntity<ApiResponse<OutletDto>> onboardOutlet(@PathVariable UUID brandId, @Valid @RequestBody OutletOnboardRequest request) {
        Outlet outlet = onboardingService.onboardOutlet(brandId, request.getName(), request.getFssaiLicenseNumber(), request.getLat(), request.getLng(), request.getTimings(), request.getBannerUrl(), request.getCuisine(), request.getRating(), request.getReviewsCount(), request.getDeliveryTime(), request.getDeliveryFee(), request.getTags(), request.getTimeZone(), request.getCityId());
        return ResponseEntity.ok(ApiResponse.success(toOutletDto(outlet), "Outlet onboarded successfully"));
    }

    @PutMapping("/api/v1/outlets/{outletId}/timings")
    @PreAuthorize("hasRole('RESTAURANT') and @restaurantAccess.onOutlet(#outletId, authentication, T(com.fooddelivery.common.enums.OrganisationPermission).OUTLET_MANAGE)")
    public ResponseEntity<ApiResponse<Void>> updateOutletTimings(@PathVariable UUID outletId, @Valid @RequestBody OutletTimingsUpdateRequest request) {
        onboardingService.updateOutletTimings(outletId, request.getTimings());
        return ResponseEntity.ok(ApiResponse.success(null, "Outlet timings updated successfully"));
    }

    @GetMapping("/api/v1/brands/{brandId}/outlets")
    @PreAuthorize("hasRole('RESTAURANT') and @restaurantAccess.onBrand(#brandId, authentication, T(com.fooddelivery.common.enums.OrganisationPermission).ORG_VIEW)")
    public ResponseEntity<ApiResponse<List<OutletDto>>> getOutletsByBrand(@PathVariable UUID brandId) {
        List<OutletDto> dtos = onboardingService.getOutletsByBrand(brandId).stream().map(this::toOutletDto).collect(java.util.stream.Collectors.toList());
        return ResponseEntity.ok(ApiResponse.success(dtos, "Fetched outlets"));
    }

    @PutMapping("/api/v1/outlets/{outletId}/status")
    @PreAuthorize("hasRole('RESTAURANT') and @restaurantAccess.onOutlet(#outletId, authentication, T(com.fooddelivery.common.enums.OrganisationPermission).OUTLET_MANAGE)")
    public ResponseEntity<ApiResponse<Void>> updateOutletStatus(@PathVariable UUID outletId, @Valid @RequestBody OutletStatusUpdateRequest request) {
        onboardingService.updateOutletStatus(outletId, request.getIsActive());
        return ResponseEntity.ok(ApiResponse.success(null, "Outlet status updated successfully"));
    }

    @PutMapping("/api/v1/outlets/{outletId}/settings")
    @PreAuthorize("hasRole('RESTAURANT') and @restaurantAccess.onOutlet(#outletId, authentication, T(com.fooddelivery.common.enums.OrganisationPermission).OUTLET_MANAGE)")
    public ResponseEntity<ApiResponse<Void>> updateOutletSettings(@PathVariable UUID outletId, @Valid @RequestBody OutletSettingsUpdateRequest request) {
        onboardingService.updateOutletSettings(outletId, request.getDefaultPrepTimeSeconds());
        return ResponseEntity.ok(ApiResponse.success(null, "Outlet settings updated successfully"));
    }

    // Legacy backwards compatibility: CustomerApp uses /api/v1/restaurants/{id}
    @GetMapping("/api/v1/restaurants/{id}")
    @PreAuthorize("permitAll()")
    public ResponseEntity<ApiResponse<NearbyRestaurantDTO>> getRestaurant(@PathVariable UUID id) {
        Outlet outlet = onboardingService.getOutletById(id);
        boolean isOpen = com.fooddelivery.restaurant.service.OpeningHours.isOpen(java.time.Instant.now(), outlet.getTimeZone(),
                outlet.getTimings(), com.fooddelivery.restaurant.entity.OutletTiming::getOpeningTime,
                com.fooddelivery.restaurant.entity.OutletTiming::getClosingTime);
        Brand brand = onboardingService.getBrandById(outlet.getBrandId());
        NearbyRestaurantDTO dto = NearbyRestaurantDTO.builder()
                .id(outlet.getId())
                .name(outlet.getName())
                .isActive(outlet.getIsActive())
                .defaultPrepTimeSeconds(outlet.getDefaultPrepTimeSeconds() != null ? outlet.getDefaultPrepTimeSeconds() : 900)
                .isOpen(isOpen)
                .lat(outlet.getLocation() != null ? outlet.getLocation().getY() : null)
                .lng(outlet.getLocation() != null ? outlet.getLocation().getX() : null)
                .image(outlet.getBannerUrl())
                .deliveryFee(outlet.getDeliveryFee() != null ? outlet.getDeliveryFee().doubleValue() : 0.0)
                .logoUrl(brand.getLogoUrl())
                .build();
        return ResponseEntity.ok(ApiResponse.success(dto, "Restaurant fetched successfully"));
    }

    @GetMapping("/api/v1/restaurants/nearby")
    @PreAuthorize("permitAll()")
    public ResponseEntity<ApiResponse<List<NearbyRestaurantDTO>>> getNearbyRestaurants(@RequestParam double lat, @RequestParam double lng, @RequestParam(defaultValue = "5.0") double radius) {
        List<Outlet> nearbyOutlets = onboardingService.getNearbyOutlets(lat, lng, radius);
        List<NearbyRestaurantDTO> responseList = new java.util.ArrayList<>();
        Map<UUID, Brand> brandCache = new HashMap<>();
        for (Outlet outlet : nearbyOutlets) {
            NearbyRestaurantDTO dto = NearbyRestaurantDTO.builder()
                .id(outlet.getId())
                .name(outlet.getName())
                .isActive(outlet.getIsActive())
                .defaultPrepTimeSeconds(outlet.getDefaultPrepTimeSeconds() != null ? outlet.getDefaultPrepTimeSeconds() : 900)
                .isOpen(true)
                .build();
            double distance = 1.5;
            if (outlet.getLocation() != null) {
                dto.setLat(outlet.getLocation().getY());
                dto.setLng(outlet.getLocation().getX());
                // approximate distance
                double rEarth = 6371.0;
                double dLat = Math.toRadians(outlet.getLocation().getY() - lat);
                double dLon = Math.toRadians(outlet.getLocation().getX() - lng);
                double a = Math.sin(dLat / 2) * Math.sin(dLat / 2) + Math.cos(Math.toRadians(lat)) * Math.cos(Math.toRadians(outlet.getLocation().getY())) * Math.sin(dLon / 2) * Math.sin(dLon / 2);
                double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
                distance = rEarth * c;
                dto.setDistance(Math.round(distance * 10.0) / 10.0);
            } else {
                dto.setDistance(1.5);
            }
            // Use real DB values, fallback to dummies if null (for old records)
            dto.setImage(outlet.getBannerUrl() != null ? outlet.getBannerUrl() : "https://images.unsplash.com/photo-1517248135467-4c7edcad34c4?auto=format&fit=crop&w=800&q=80");
            dto.setCuisine(outlet.getCuisine() != null ? outlet.getCuisine() : "Multi Cuisine");
            dto.setRating(outlet.getRating() != null ? outlet.getRating() : 0.0);
            dto.setReviewsCount(outlet.getReviewsCount() != null ? outlet.getReviewsCount() : 0);
            dto.setDeliveryTime(outlet.getDeliveryTime() != null ? outlet.getDeliveryTime() : 30);
            dto.setDeliveryFee(outlet.getDeliveryFee() != null ? outlet.getDeliveryFee().doubleValue() : 0.0);
            if (outlet.getTags() != null && !outlet.getTags().isEmpty()) {
                dto.setTags(java.util.Arrays.asList(outlet.getTags().split(",")));
            } else {
                dto.setTags(List.of());
            }
            try {
                Brand brand = brandCache.computeIfAbsent(outlet.getBrandId(), id -> onboardingService.getBrandById(id));
                dto.setLogoUrl(brand.getLogoUrl());
                dto.setBrandId(brand.getId());
                dto.setBrandName(brand.getName());
                if (outlet.getBannerUrl() == null && brand.getLogoUrl() != null) {
                    dto.setImage(brand.getLogoUrl());
                }
            } catch (Exception e) {
                log.debug("Brand not found for outlet: {}", outlet.getId());
            }
            // Ignore missing brand
            responseList.add(dto);
        }
        // Sort by distance (since DB groups by brand, we sort the final list by distance)
        responseList.sort(java.util.Comparator.comparingDouble(NearbyRestaurantDTO::getDistance));
        return ResponseEntity.ok(ApiResponse.success(responseList, "Nearby restaurants fetched"));
    }

    @GetMapping("/api/v1/restaurants/brands/{brandId}/outlets")
    @PreAuthorize("permitAll()")
    public ResponseEntity<ApiResponse<List<NearbyRestaurantDTO>>> getBrandOutlets(@PathVariable UUID brandId, @RequestParam double lat, @RequestParam double lng, @RequestParam(defaultValue = "5.0") double radius) {
        List<Outlet> nearbyOutlets = onboardingService.getNearbyOutletsByBrand(brandId, lat, lng, radius);
        List<NearbyRestaurantDTO> responseList = 
        nearbyOutlets.stream().map(outlet -> {
            NearbyRestaurantDTO dto = NearbyRestaurantDTO.builder()
                .id(outlet.getId())
                .name(outlet.getName())
                .isActive(outlet.getIsActive())
                .defaultPrepTimeSeconds(outlet.getDefaultPrepTimeSeconds() != null ? outlet.getDefaultPrepTimeSeconds() : 900)
                .isOpen(true)
                .build();
            double distance = 1.5;
            if (outlet.getLocation() != null) {
                dto.setLat(outlet.getLocation().getY());
                dto.setLng(outlet.getLocation().getX());
                double rEarth = 6371.0;
                double dLat = Math.toRadians(outlet.getLocation().getY() - lat);
                double dLon = Math.toRadians(outlet.getLocation().getX() - lng);
                double a = Math.sin(dLat / 2) * Math.sin(dLat / 2) + Math.cos(Math.toRadians(lat)) * Math.cos(Math.toRadians(outlet.getLocation().getY())) * Math.sin(dLon / 2) * Math.sin(dLon / 2);
                double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
                distance = rEarth * c;
                dto.setDistance(Math.round(distance * 10.0) / 10.0);
            } else {
                dto.setDistance(1.5);
            }
            dto.setImage(outlet.getBannerUrl() != null ? outlet.getBannerUrl() : "https://images.unsplash.com/photo-1517248135467-4c7edcad34c4?auto=format&fit=crop&w=800&q=80");
            dto.setCuisine(outlet.getCuisine() != null ? outlet.getCuisine() : "Multi Cuisine");
            dto.setRating(outlet.getRating() != null ? outlet.getRating() : 0.0);
            dto.setReviewsCount(outlet.getReviewsCount() != null ? outlet.getReviewsCount() : 0);
            dto.setDeliveryTime(outlet.getDeliveryTime() != null ? outlet.getDeliveryTime() : 30);
            dto.setDeliveryFee(outlet.getDeliveryFee() != null ? outlet.getDeliveryFee().doubleValue() : 0.0);
            if (outlet.getTags() != null && !outlet.getTags().isEmpty()) {
                dto.setTags(java.util.Arrays.asList(outlet.getTags().split(",")));
            } else {
                dto.setTags(List.of());
            }
            try {
                Brand brand = onboardingService.getBrandById(outlet.getBrandId());
                dto.setLogoUrl(brand.getLogoUrl());
                dto.setBrandId(brand.getId());
                dto.setBrandName(brand.getName());
                if (outlet.getBannerUrl() == null && brand.getLogoUrl() != null) {
                    dto.setImage(brand.getLogoUrl());
                }
            } catch (Exception e) {
                log.debug("Brand not found for outlet: {}", outlet.getId());
            }
            return dto;
        }).sorted(java.util.Comparator.comparingDouble(NearbyRestaurantDTO::getDistance)).collect(java.util.stream.Collectors.toList());
        return ResponseEntity.ok(ApiResponse.success(responseList, "Brand outlets fetched"));
    }

    @GetMapping("/api/v1/internal/admin/restaurants/all-with-location")
    @PreAuthorize("hasAnyRole('ADMIN', 'SERVICE')")
    public ResponseEntity<ApiResponse<com.fooddelivery.common.dto.PageResponseDto<NearbyRestaurantDTO>>> getAllOutletsWithLocation(
            @RequestParam(required = false) @com.fooddelivery.common.location.CityId @Pattern(regexp = com.fooddelivery.common.location.CityIdValidator.REGEX) String cityId,
            @RequestParam(defaultValue = "0") @Min(0) int page,
            @RequestParam(defaultValue = "100") @Min(1) @Max(100) int size) {
        org.springframework.data.domain.Pageable pageable = org.springframework.data.domain.PageRequest.of(
                page, size, org.springframework.data.domain.Sort.by("id").ascending());
        org.springframework.data.domain.Page<Outlet> outlets = onboardingService.getAllOutletsForFleetCity(cityId, pageable);
        
        // A missing point is not a location at (0,0). Keep the scoped source-page metadata so
        // callers can continue through pages, but omit the unsafe fake pin from the map payload.
        java.util.List<NearbyRestaurantDTO> locatedOutlets = outlets.getContent().stream()
                .filter(outlet -> outlet.getLocation() != null)
                .map(outlet -> NearbyRestaurantDTO.builder()
                        .id(outlet.getId())
                        .name(outlet.getName())
                        .isActive(outlet.getIsActive())
                        .cityId(outlet.getCityId())
                        .lat(outlet.getLocation().getY())
                        .lng(outlet.getLocation().getX())
                        .build())
                .toList();
        // PageImpl normalizes a total down to the filtered content size when page size exceeds
        // the total. Keep the source pagination metadata so callers can visit later source pages,
        // while reporting the actual number of safe, drawable pins in this response.
        com.fooddelivery.common.dto.PageResponseDto<NearbyRestaurantDTO> responsePage =
                com.fooddelivery.common.dto.PageResponseDto.<NearbyRestaurantDTO>builder()
                        .content(locatedOutlets)
                        .totalElements(outlets.getTotalElements())
                        .totalPages(outlets.getTotalPages())
                        .last(outlets.isLast())
                        .size(outlets.getSize())
                        .number(outlets.getNumber())
                        .first(outlets.isFirst())
                        .numberOfElements(locatedOutlets.size())
                        .empty(locatedOutlets.isEmpty())
                        .build();
        return ResponseEntity.ok(ApiResponse.success(responsePage, "All restaurants with locations fetched"));
    }
}
// @Getter